Use a quadratic curve instead of an arc
Using an arc was given too much problems. Basically because
of transforming the radius, the stroke weight for x and y
for each arc using the transformation matrix.
We basically use a better approach which is using a quadratic
curve withe the control point set at the intersection point
of the two edges.
Also fix a bug we had by using the first and last vertex for the
last edge to process which is wrong, we should pass the first edge.