Only check for implementation changes
Given that the shap path is a shape, all the shape
attrbiutes are already checked for change. In case any change
happens on the shape path renderer it is automatically propagated
to the path renderer inside the shape path renderer and thus
it will make a change when there's no need for it.