Fix a buffer overflow
Basically the algoirthm checks for all the edges that intersect
an particular y and gets the rightmost and leftmost coordinates.
Then it tries to clip such coordinates to the input x,y, len triplet.
Right now the algorithm was not checking for the case where
the x is right to rx, that is, it is beyond the actual edge
intersection. For such case we need to memset with zero too.