Add API to set/get the buffer/surface alpha hints
This might not impact how an input surface should change its
own alpha hints based on the rendering of a renderer. The reason
is that a renderer might draw only on an area, and if such drawn
area is alpha full, still the surface might have a non full alpha
in a place it was not drawn.

We better make this internal only later