Improve the opencl setup
We no longer use the source_name as kernel name. Let the renderer
define more than one kernel and be itself the one who provides
the name of the kernel explictly.

The surface is added as a kernel arg in the lsat moment before
rendering instead of the setup. This is because inner renderers
do the setup with the parent's surface but they are drawn into
another surface