Cache the programs binaries into the filesystem
In order to improve the setup speed of the first rendering
instead of compiling everytime the same OpenCL code for a
particular device we dump the binary associated with a
program into the filesystem. Then we just load the binary
and create a program from that. The speed up is huge :)